More efficient handling of patterns in exact hints.
We know that the pattern for exact hints, in discriminated mode, is always
set to the type of the hint. Instead of turning it into a constr_pattern stored at
runtime and then turned into a dnet pattern, we treat this specifically and
rebuild the dnet pattern directly out of the hint type.

This should be both more efficient and less memory-hungry.