Do not nf-βι normalize case-related error messages in Type_errors.
This was the only place in the kernel where the strong normal form functions
were used. Furthermore, there are various reasons why it shouldn't be done.
First, this is redundant, it is already performed by the printer. Second,
the EConstr equivalents in Pretype_error do not perform this reduction.
Finally, since the change of case representation there is no point in trying
to mangle the branches to produce something that looks like what the user wrote.