Extend the -async-proofs-tac-j option to accept the value 0.
When this option is passed the integer 0, it now replaces the par: selector
with an equivalent combinator that does not spawn any new process and behaves
as all: solve [tac] instead. This is different from when passed the integer 1,
which means that a process is still spawned although it doesn't parallelize
anything.