Fix representation of empty arrays (fix #16832).
Since there are no arguments, Vmbytegen.comp_args cannot be called in that
case, so we just skip it.

The MAKEBLOCK opcode needs to be adapted accordingly, since size-0 blocks
have a special representation in OCaml.

Remark: OCaml's bytecode interpreter uses special opcodes ATOM, PUSHATOM,
ATOM0, and PUSHATOM0, for that purpose; but that seems excessive in the
case of Coq.