Add a staging notion to option declarations
In rare cases, options can have an impact on parsing, hence have to be
computed in the synterp stage. The `Default Proof Mode` option is such
an example.

This is a step towards #15409.

When interpretation is separated into synterp and interp, Set commands
will call set_option_value once in each stage, and set_option_value
will ignore the call in the non-matching stage.