Internalizing a qualid passed to an Ltac as a constr with its impargs.
This was the case in non-strict mode (via interpretation at runtime),
but, in strict mode, it was instead the reference without its maximal
implicit arguments.

The main impact is on "class_apply" which we reformulate as a tactic
notation to ensure that its argument is always interpreted as a term.