[parsing] Remove hack which hampered correct parsing resumption.
This is a follow up of #14220 and #14898 ; in particular, we revert
the hack in `CLexer.func` and instead require the caller to set the
stream to right count, as this is critical for the current
architecture.

Together with #16476, which addedd `Parsable.loc`, and the export of
`CLexer.after` in this PR, it seems we can now properly implement
parsing resumption (and memoization) in the following manner:

- when done with parsing, use `Pcoq.Parsable.loc` to get the last
  lexing position `loc` used by the parser.
- advance `loc` using `CLexer.after`
- setup a new stream with the resuming text, you *must* make sure the
  `count` in the stream corresponds to `loc.bp`, easiest is to use
  `Stream.njunk`.
- pass `loc` to `Pcoq.Parsable.make ~loc`

Memoization is a bit more advanced and requires using
`Pcoq.Parsable.consume`.

It would be nice drop the `Stream.njunk` requirement, but Coq's lexer
internals do heavily rely on `Stream.count` so that would imply a full
rewrite of the lexer which is something I'd rather don't do now.

Unit test added.

Fixes #15087

(cherry picked from commit 5898d523d8f87863fd911a0d6b7a0be99e5b1700)