More compact representation of VM patch positions.
We exploit the fact that positions are guaranteed to fit into at most 32-bit,
and that the relative distance between positions is typically small. Instead
of using an array of integers, we use a variable-length encoding of integers
coded as a string. We also replace the integers by the diff with the previous
one.