[coqide] Cleanup location code, in particular byte to offset handling.
This should make the logic a bit more clear; it is mostly a cleanup,
as the locations sent to the server should be the same.

Note that we can use GTK line-cache to directly convert a Coq-style
location to GTK iterators, so we do so and drop the `byte_start` part
from the sentence structure.