[extraction] Rename `Pervasives.op` -> `Stdlib.Int.op`
This is required for OCaml 5.0 support, as `Pervasives` is gone.

The bound on extraction was already OCaml 4.07 so I think this is fine
as is. Note that `Stdlib.Int.{max,min}` only appear in 4.13 so I've
mapped them for now to the polymorphic variants.

Fixes #11359

(cherry picked from commit 54e1ddd0a75e5ea980aec55510d5befb242c787d)