[clib] [cThread] Don't inline `Mutex.unlock` manually.
This was added as an optimization in #14046 , however it is not
portable to OCaml 5, due to a renaming.

Instead of introducing some conditional compilation, we rely on the
`inline` and `inlined` attributes, which should do the same job, and
given that this is a very small optimization feels like less work.

(cherry picked from commit bfc6148c116b113408c90c05e224a79b440c123c)