[docker] Bump some package versions for improved forward compat.
We mostly cherry pick from #15494 , so `CI_OPAM` will look more
similar to what would be needed in a "5.0" switch.

The only important bump is `dune` for the edge switch, which will be
useful to test some projects that require a higher Dune version in
order to build (c.f. #16964)