[loc] Make location type private, start to enforce some invariant.
Due to historical reasons, Coq used the location type present in
Camlp5, and some cruft did accumulate in the lexer as the bugfixes
couldn't be usptreamed easily.

This is not the case anymore, so we make the location type private and
start to enforce some invariants.

Moreover, some parts of the code rely on old-style [start-end] pure
offset-based location. While there is nothing wrong with this, it is
extremely convenient for IDEs that Coq provides both the offset and
the line / column number (so for example IDEs can reuse editors line
to offset conversion cache to handle unicode etc...)

We thus remove the APIs that allow to create locations with missing /
incorrect line numbers.

c.f. #17031

TODO:
 - Remove Loc.after
 - Remove Loc.make_loc