Refinement when printing recursive notations where "forall" binds Anonymous.
Previously, we expected a "forall" in a recursive notation to not bind
Anonymous, so that the notation "A -> B" is systematically
preferred, as e.g. in:

Notation "∀ x .. y , P":= (forall x, .. (forall y, P) ..)  (x binder, at level 200, right associativity).
Check nat -> nat.
(* "nat -> nat" and not "∀ _ : nat, nat" *)

The refinement is to give priority to "A -> B" only when the
corresponding "forall"s are nested on the right. For instance, with
the commit, the heuristic does not apply any longer in:

Notation "!! x .. y # A #" := (.. (A,(forall x, True)) ..,(forall y, True)) (at level 200, x binder).