Fix backtracking over plugin vernac extensions
Fix #12575

This was hand tested using
~~~coq
Require Import Ltac2.Ltac2.

Ltac2 Eval ().
~~~
we must correctly handle
- eval all, backtrack before the Require, eval just Ltac2 Eval: should
  parsing error illegal begin etc
- eval all, backtrack before the Require, eval all again: should
  work (ie the parsing rules must be re-applied)

The interesting part is what happens in vernac_extend and the
associated egramml functions:
- an indirection is added from the command name (Foo in
  `VERNAC COMMAND EXTEND Foo`) to the extension data (which contains a
  closure: not marshallable).
  This is non-backtracking, and there is no use to making it backtrack
  since we can't unlink plugins.
- vernac_extend calls `Egramml.extend_vernac_command` with `undoable:false`
  when there is no `plugin` argument, this is the old non
  backtrackable behaviour. Typically for non-plugin callers like
  g_obligations.
  Eventually we will clean up so that rules added this way produce an
  error when backtracked over instead of having the redo code in pcoq,
  until then it's also available for plugin hacks which don't mind the
  backtrack bug I guess.
- when there is a `plugin` argument, vernac_extend wraps the call to
  `Egramml.extend_vernac_command` in `Mltop.add_init_function` and
  uses `undoable:true`. In undoable mode, egramml uses the
  Pcoq.create_grammar_command system which backtracks properly.
  Because this system puts its input in the grammar_stack which ends
  up in the summary, we need the indirection mentioned above to keep
  the input marshallable.

For GRAMMAR EXTEND there is no user-given name so we produce a uid
from DECLARE PLUGIN, the file name and the number of occurences of
GRAMMAR EXTEND in the file (we end up with eg
`coq.plugins.ltac:coretactics.mlg:42`).

In the future we will probably try to replace the redo code of
ByEXTEND with an error and make the users in plugins switch to the new
mechanism.
For now it is too much work, when a plugin is modifying its own
entries it doesn't matter if backtrack is undone.