Mostly functionalize CLexer, move state ref to pcoq
BEWARE CLexer.Lexer.tok_func and Grammar.GMake.Parsable.make take a
`keyword_state ref`!

Indeed when parsing a file we do `Parsable.make` on the `Stream` from
the file and the resulting `LStream` needs to somehow get updated when
the keywords change.

This will require changes to the Stream/LStream APIs to address.