Simplify the implementation of the decompose tactic.
The code was pretending to handle the case where a non-record inductive term
being decomposed appeared in the goal, but in practice this was not working
at all (it looped or produced garbage) and probably never worked. Instead of
calling the recursor we thus simply reuse the code as the one for dependent
case analysis.