Remove all uses of the functional induction tactic from the stdlib.
This reduces the reliance of the stdlib on the plugin whose name must not
be uttered. The diff is so trivial I do not really understand why we have
a spaghetti wrapper around induction except for adding fragility. It should
be the job of induction to perform the administrative task of finding the
parameters.