Fixes #17073: anomaly with inaccessible branchs of "match" in guard checking
Indeed, do not expect an inaccessible "match" branch to have a type
compatible with the external type. So, if there is a stack of
arguments to apply to the whole "match", the inaccessible branch is
not necessarily able to grab them.

(cherry picked from commit 47962163d7325db976f9c718598f83be59eed168)