Correctly compute sort relevance in Retyping.
Instead of relying on the sort family, which lacks the necessary data
to return a variable relevance, we extract directly the relevance from the
corresponding sort.

We should probably review seriously all calls to Sorts.relevance_of_sort_family,
but for the time being let us do this piecewise.

Fixes #17124.