Partially fixing #17137 on coercions missing when matching in singleton types.
This is only a partial fix, because there is a test of irrefutability
of a line of patterns which is incorrect when coercions are liable to
be inserted.

The original test of irrefutability was uselessly too permissive. The
test of irrefutability only matters when the first pattern is a
variable (so that the split on this variable can eventually be
skipped), but it was done also in the non-variable case, disregarding
that in the non-variable case a coercion can change the singleton
status of the pattern.