[time] [coqc] Print time for document setup and end
This can be useful when benchmarking.

As header in the time output, we use a special marker with a Coq
comment so it is still valid .v syntax: "(* Document Start / End *)"

Note that we remove the "LoadingPrelude" phase, otherwise we get
strange output.

IMHO that's not very useful, and indeed I hope one day we can have a
model where no prelude is loaded automatically.

Co-authored-by: Gaëtan Gilbert <gaetan.gilbert@skyskimmer.net>