[declare] API for obligation display from Users Interfaces
As of today, the only interface to explore obligations is
`show_obligations` which is far from optimal for anything not
display-based.

We introduce a preliminary view on the obligation state, that UI can
use.

This has been tested on https://github.com/ejgallego/coq-lsp/pull/262

Next commits will push the toplevel-specific `show_obligations` code
to the command interpretation.