Move the evar clause API out of Clenv.
Apart from the evar clause API, the Clenv module only contains morally deprecated
functions that ought to be removed at some point. Thus it is cleaner to expose
the only relevant part in a dedicated module. We put this module in Tactics
since proofs/ seems to be very ill-fitted for this purpose.