During substitution, do not discard canonical part of inlinable constants
Fixes #17261

This attempts to revert an old fix for #2608, which does not appear to be
correct:
https://github.com/coq/coq/commit/63710768d32c5929aa6498aff5aae2f9e0058e9
The problem that occurs in #2608 no longer seems to exist because the code has
been substantially rewritten, as can be seen from the succeeding test case in
that commit.

I'm not entirely sure that what I'm doing here is correct. Somebody
knowledgeable about the module system should take a closer look.