[configure] Fix "local layout" install for Windows
These are some leftovers of the "local" layout that went away some
time ago. But in Windows, since 8.17 we will generate bad paths for
example for `coqlib`, as Dune will install in `C:\coq\lib\coq` not on
`C:\coq\lib` like before.

We could special case these, but IMHO not a good thing.

This bugs was not observed because opam packages then to set
`--libdir` manually anyways.

Note that still users have to be careful to pass the same install
paths to configure and to `dune install`