[`setoid_rewrite`] rewrite hypotheses without evar instance shenanigans
Fix #17295

The previous approach of rewriting a hypothesis by putting the proof term for
the new hypothesis in the instance of the new evar seems fundamentally broken.
Because it replaces the old hypothesis with a new one, subgoals no longer have
access to the old hypothesis.

Instead, we use the conventional method of generating a new hypothesis, and then
replacing the old hypothesis with the new one, clearing any dependent
variables.

Also constitutes a nice cleanup of rather suspicious looking code.