Extraction to OCaml: use correct key to manage collision of record field names.
Fixes #3846
Fixes #14843
Fixes #16677
Fixes #12813

When dealing with collisions by means of encapsulation in a Coq_XX
module, a field name should be bound to the name of the corresponding
inductive type declaration.

We use Table.repr_of_r to do that, but early enough to have the
relevant information on how to bind a field to the name of the
inductive that declares it. (In particular, we cannot rely on the
global tables in structures.ml because they are not valid when inside
a functor, see #17321.)

Note that Haskell and Scheme do not use records and it is ok on their
side.

(cherry picked from commit c19b2cd5d2ee8df9897187bab0b96947ad666ae7)