[library] Make require more independent of files.
This is a step towards allowing in-memory requires.

We don't export the right API yet as I think we should do some further
refactoring internally in `library.ml`

There is a minimal change of behavior in terms of feedback, before we
only emitted the file location for the first module, we now do for
all. That should be harmless (and in fact maybe better as the file
access is really happening)