[library] Allow require to load libraries from memory.
This is one more step in our quest to avoid saving / loading .vo
files to disk when not needed.

Most changes are straightforward except for the extension of the
command type to include an internalization function.

For example, a typical use case in jsCoq is to download a .vo file on
Require, but there are many more possibilities.