[vernac] Remove LoadPath edition vernacular commands
Deprecated since 8.16, in general we prefer setting loadpaths outside
of Coq documents.

This will allow us to provide to vernac mostly a read-only interface
to the FS. We could indeed allow the documents to update this kind of
state by setting the right write callbacks, however IMHO that's not
worth it.