Export a variant case-building function tailored for case_pf.
This might marginally change the autogenerated names in destruct because
the environment used to compute them is slightly different. Branches are
now independent, so if ever the autogenerated name for a branch triggered
a rename for an argument, it wouldn't happen anymore now. It is not even
clear this really matters since case_pf takes care of renaming branch
names anyways, only internal proof terms might differ.