Fix a discrepancy in the branch variables and introduce hypotheses in case_pf.
The old code was generating a fresh set of variables just for the branches
from the internal Case node, but relied on another set of fresh names for the
type of the branch. The latter used to be more precise because it had access to
the parameters whereas the former did not.

Since the tactic layer only cares about the variables from the type, this
discrepancy only affected printing of pattern-matching.

Now that we build the branch with the parameters already instantiated, this
code is redundant, which is why we remove it in this commit.