Explicitly passes the parameters to the case generating function.
This reduces the amount of trivial substitutions we have to perform. This might
also slightly change the generated names for the non-dependent variables bound
in the return clause, since their type might have changed now. Nonetheless
this should be an extremely fringe difference, furthermore only mattering for
error messages and case printing.

Actually the test-suite does catch such an example.