Stop reducing case analyses when the head is a constructor in case_pf.
This is a very weird corner case of the case tactics and its variants. If
the argument βι-normalizes to a constructor, the case node will disappear
as it is also reduced by the same reduction strategy. This is specific to
case-like tactics because destruct and friends internally only apply to
Var arguments which are left untouched by βι-reduction.

After this patch we treat the argument in a fully rigid way, so that a call
to e.g. case 0 with generate two subgoals and not only one. I doubt that it
really happens in the wild, barring some arcane design pattern I have not
forecast.