Add CMap.ExtS.symmetric_diff
`symmetric_diff` efficiently calculates the difference between two maps when
they share most of their internal structure.

Note that this cannot be implemented on top of the public interface of CMap, as
it relies on the internal implementation details. Tactician now accesses the
Coq internals through `Obj.magic`, and Coq in turn accesses the OCaml internals
through `Obj.magic`, which seems somewhat risky. So I'm hoping that this code
can find a home here. But eventually I feel that the whole `CMap` extension to
OCaml's map should be upstreamed there...