Refman debug mode
When COQ_DEBUG_REFMAN is defined in the environment, duplicate
everything we send to coqtop to a file /tmp/coqdomainXXXX.v

This spams many files so not enabled by default.

Also disable the antlr version check when the environment variable is
set as I couldn't figure out how to install the right version on my
machine (it seems to work fine with 4.9.1 instead of 4.7.2).

The temporary file is put explicitly in /tmp instead of the automatic
tmpdir as otherwise it goes in some /tmp/ subdirectory created by dune
and deleted by dune. If we can disable the deletion that could be nicer.