Fix inverted logic in `Summary.ref ~local:true`
We want to keep the value when `marshallable` is `false` not `true`.

Fixing the logic reveals a bug where the state sent to proof workers
was produced with marshallable:false, so instead we redesign the API.

Now freeze_function is unit -> 'a, and summaries may be declared with
an optional make_marshallable:'a -> 'a.