Improve unable to satisfy constraint message when exactly 1 unsolvable
This is a guess at what the error_unresolvable code was trying to do.
Previously `found` was always `true` so `ev` was always `None` and we
folded for nothing.

This was the case since 73c5ecd3d038b4143910762c0132e147c56a85a2 which
changed the initial fold value from `false` to `true`.

We could go further and pick some evar to emphasize when there are
multiple, but not clear how to pick.