Do not create bytecode aliases for primitive operations (fix #17663).
When dynamically linking the bytecode from an instantiated functor, the
kernel expects all the symbols to have some associated bytecode (or to be
axioms), so that it can relocate them. Aliases to primitive symbols,
however, have no bytecode and were thus wrongfully considered as axioms.

This commit forces the creation of bytecode in place of aliases for
primitive operations.