[coqide] Improve offset compensation
Turns out that we need to be a bit more careful at mark addition
points as sometimes our info is coming from buffer's view and other
times location information is coming from Coq which has an outdated
view.

For the tooltips, we make the choice to store them in Coq locations,
which seems like a reasonable ground truth, as in the IDE side things
could be much more dynamic, but it is also very efficient to perform
the correction as the GTK iterators have the needed info in constant
time.

We also need to compute the drift of the line for the UTF-8 conversion
fiction, otherwise we get things wrong on line-drifts.