Split Docker images in two: one for base and one for edge.
The base image will use the old Ubuntu LTS (currently 20.04) and the
edge image will use the most recent Ubuntu version (currently 23.04).

We temporarily switch the dev job to use the base image because the
documentation jobs depend on it, but should be tested with our minimal
dependency version bounds.