Moving ARGUMENT EXTEND rewstrategy grammar to GRAM EXTEND to enforce assoc.
Note that the original grammar is preserved but I don't know how to
present it in a way stratified with levels and unique associativity
per level.

We also seize this opportunity to rewrite the printer (also in an ad
hoc manner) so as to respect the grammar.

Example of non-level-stratified behavior:

 subterms id ; id ; choice id ; id ; subterms choice id ; id ; id

is interpreted as

 ((subterms id) ; id) ; (choice ((id ; id) ; subterms (choice ((id ; id) ; id))))

that is:
- "subterms" (and all prefix keywords other than "choice") bind narrower than ";"
- "choice" binds wider than ";"
- ";" is left-associative

To get "choice" binding wider than ";" we need it to be at a higher or
equal level.

If ";" were right-associative, we could have "choice" and ";" both at
same top level with right associativivity.

But ";" is left-associative, so, in "s ; choice s'", "choice" has to
be at a lower level.

Then, "choice" has to take an argument from a higher level than where
it is. This is visible in the printer which has to remember when
"choice" is used in a context which overlaps with this higher level.

Co-Authored-By: Théo Zimmermann <theo.zimmi@gmail.com>