[coqidetop] Fix print depth setting for Richpp clients
cc: #11710 #13971

This doesn't solve the bug above in full, as some clients like CoqIDE
do client-side rendering.

These need to either query Coq for the depth setting at a particular
doc position, or use their own, client-side setting.

Note that the problem of querying Coq for depth setting is that it can
vary for particular document points, and it is not the case that IDE
queries do necessarily refer to a particular doc point.