Add printer support for constr rules of the form "SELF ; ... ; term LEVEL XX".
That is, we want to revert situations of the form:

term:
[ "XX" [ SELF ; "op"; ... -> ... ]
| "YY" [ "prefix" ; term LEVEL "XX" -> ... ] ]

where "prefix ... op ..." is parsed as "prefix (... op ...)" and we
want "(prefix ...) op ..." not to be printed "prefix ... op ..." as of
today, which is wrong, but really as "(prefix ...) op ...".

This is done by remembering the level at which the continuation of a
printing is done. That is, when printing "prefix ...", we remember
that there is a continuation "op ..." at level "XX". This allows to
know that parentheses are mandatory on the left to prevent the
reparsing of the printing to absorb the continuation into "prefix
(... op ...)" instead of the expected "(prefix ...) op ...".

Then, whenever we print a rule "... ; SELF LEVEL XX" and there is a
continuation at level less or equal than XX (in
may_capture_cont_after), we:
- add parentheses around the whole notation
- we cancel the existence of a continuation on the right since the
  collision is protected (in adjust_level)

Also:
- in each rule of the form "SELF; sep; SELF; sep ..." at level XX, we set
  the continuation in the left-hand side to be at level XX
  and the inner SELF to no continuation
- in each rule of the form "SELF; SELF; SELF; ..." at level XX, we set
  the continuation to be at level XX for all SELF.

Also, this concerns only constr notation: custom notations are treated
in constrextern.ml.