For arguments scopes, statically know the list of available scopes.
This allows to rebuild the scopes knowing only the discharged data
while, before, an argument scope at some level dynamically depended on
which pairs of a scope binding a reference were registered in the
current section level.

Also make more local the trick of passing an integer from discharge to
rebuild.