Fixes #11237: support constructor parameters in prim notations for patterns.
Additionally, we restructurate the code a bit, moving the check on
well-formedness of patterns in constrintern.ml so that the separation
of roles between check_allowed_ref_in_pat (just check the validity so
that another interpretation can be taken if it fails) and rcp_of_pat
(which interpret the pattern) is clearer.

We also check that the "in" clause of "match" accepts a notation for
inductive type at the head.