Pass bound universes directly as an instance in the VM.
This makes the VM compilation scheme of bound universe variables closer to
the one of the native compiler. It also reduces the amount of free
variables in closures containing bound universe levels.

We also remove a suspect but not outright incorrect piece of code in Vconv.
For polymorphic constants with a non-empty instance, the previous code used
to call VM conversion on the universe levels of the instances. This is fine
because 1. opaque constants are invariant for instance cumulativity and 2.
the level representation is a first-order datatype for which structural
equality and semantic equality coincide, so VM conversion behaves the same
as Level.equal. Nonetheless, this is a very fragile piece of code, so we
replace it with the same code used for inductive instance comparison with
the flex flag switched on.