Ltac2: add type annotation to generated code for constr/goal_matching
We also annotate the core definitions in Pattern.v, this should
keep the type definition in sync with its usage.

With this and the previous bidi commit, #17869 produces a precise type error.

The annotation is a bit adhoc (only these 2 scopes get it), not
sure what the best API is. Maybe Tac2core.add_expr_scope should take
an additional raw_typexpr argument and it would get threaded around
until CTacSyn is turned into CTacLet and typechecked?