Update on #15582: restore decompose_lambda_n_assum in Term as it is in EConstr.
Issue #15562 questioned the arbitrariness of what to do with trailing
LetIn when decomposing a block of binders by counting only the Lambda.
Following this, commit c2e7e23f8a in PR #15582 changed the semantics
of Term.decompose_lambda_n_assum (formerly Term.decompose_lam_n_assum)
to not collect LetIn.

This introduced a discrepancy with EConstr.decompose_lambda_n_assum
(formerly EConstr.decompose_lam_n_assum) which continued to collect
LetIn.

This also overlooked that the initial behavior for LetIn was exactly
what is needed to extract the initial context of a Fix.

So, we restore the initial behavior (even if
Term.decompose_lam_n_assum is not used in the Coq archive).