Fast path in strong simpl reduction for case / proj.
The implementation of strong reduction for simpl is extemely quirky. It
recursively applies whd_simpl on all subterms through term mapping, which
is redundant. Indeed, heads of application nodes are reduced twice, once
with full access to arguments and once with an empty stack. This patch
shortcuts the second call for case and proj nodes, since it is known
statically that they will not make further progress.

In an ideal world constants should also be handled similarly, but due to
the weird implementation of simpl some developments rely on what is
arguably a bug of simpl. The nomatch modifier makes constant simplification
not monotone w.r.t. stack shortening, see e.g. #3989.