Rename Reduction.hnf_decompose_lambda_n_decls using "_assum".
Indeed, it was preserving LetIn but not counting them, which is the
same as what EConstr.decompose_lambda_n_assum did (up to forcing
reduction), while the suffix _decls suggested counting LetIn (as in
EConstr.decompose_{prod,lambda}_n_decls and
Term.decompose_{prod,lambda}_n_decls).

The discrepancy was overlooked in #15582.