Implement resolve_tc in Ltac2 instead of Ltac1
Focusing is much easier to control in Ltac2 and the bridge to call it
from Ltac1 does not force focusing.

Currently we don't make the rest of the stdlib depend on Ltac2 so
there is no convenient place to put the Ltac1 shim. It is small so I
guess users will just copy from the test or write their tactics in Ltac2.