[CI] Add coq-neural-net-interp
Tests:
- typeclasses
- notations
- Ltac2
- primitive ints
- primitive floats
- primitive arrays

Should be a pretty quick build, in part because most of the proofs
depend on a long chain of deps including MathComp, VST, Interval, and
vcfloat, which are not all on Coq's CI yet.

If we want to test `native_compute` on the CI, we can add the `computed`
target to the list of built targets.  But this will probably add about
10+ minutes to the CI for no particularly good reason.