Efficiently track evars with candidates.
Instead of crawling the whole set of undefined evars in O(n), we keep a
set of evars with actual candidates. Since most evars do not have candidates
and furthermore we only care about the evar with candidates with highest
identifier which can be fetched in O(log n), this should be much faster.