Remove distro-specific badges from README.
As of today most of them are outdated: Chocolatey and Conda do not
seem maintained anymore, Arch is at 8.16.1.

The recommended ways to install Coq are anyway already well documented
on the download page which is linked just below.

We only keep a single Repology badge as well as those for the Docker
images.