Fix CODEOWNERS for doc/stdlib.
It never made much sense to have the code owners for this directory be
the doc-maintainers while the actual content of the stdlib
documentation is written within the .v files. In practice, this has
just meant that doc-maintainers get a review request when adding or
removing a file in the stdlib, which is a bit absurd.