Merging Mrec and Nested(NestedInd) which behave the same kernel-wise.
Outside kernel, induction schemes skip nested types, so it matters but
we compensate by checking the name of the inductive type.

The test checks that the computation of recursively uniform arguments
in indtypes.ml is done only on the main inductive (the test comes from
Iris; probably a variant with 2 nesting instead of 3 is also
possible).