Node of arrays of arrays in rtree.ml so as to save tricky encodings.
Recursive trees (rtree.ml) were used to encode the algebraic structure
of (co)inductive types. The latters have a disjunctive-conjunctive
structure which was trickily encoded using an artificial Norec node in
rtree.ml. The encoding has bad side effects, such as:
- decoding the printing of trees was difficult, having to distinguish
  the fake Norec's from the regular ones
- a function like inter was not able to know the difference between
  the two uses of Norec, making its parametrization over an equality
  of tags suboptimal

The commit hard-wire a two-level of arrays in rtree.ml so as to avoid
the encoding.

At the same time, we rework a bit the printer.

In passing, we align to the convention that pr_ means to Pp.t and pp_
means to string.

----------------------------------------------------------------------
For the record, the former internal encoding of wf_paths for the
underlying disjunctive-conjunctive of inductive types was of the form:
  Node (Mrec(name-of-ind),
       [|Node (Norec,wf_paths-of-arguments-of-constructor_1);
         ...
         Node (Norec,wf_paths-of-arguments-of-constructor_n)|]
and the internal encoding of wf_paths for non-recursive
arguments was of the form:
  Node (Norec,[||])

The encoding of the underlying inductive structure is now:
  Node (Mrec(name-of-ind),
       [|wf_paths-of-arguments-of-constructor_1;
         ...
         wf_paths-of-arguments-of-constructor_n|]