Update cardinal_Add_In proof to shorter version
This shorter version was written by Jean-Christophe Léchenet in the
feedback on #12096, as a proposed alternative to my longer and
unwieldier proof.

The only change I've made is to inline `remove_In_Add`, which is a lemma
we don't want to expose immediately. Morever, I've had to adjust the
proof slightly because a use of `subst` didn't work on my end - perhaps
this is a Coq change.

Co-authored-by: Jean-Christophe Léchenet <eponier@via.ecp.fr>