Do not use the legacy pattern_of_constr for variables from typed patterns.
This call comes from the expansion of variables bound to a term in reduction
tactics, e.g. writing something along the lines of `let x := ... in simpl x`.

I think that replacing the broken variant with the regular one should not
matter, since there is virtually no way to exploit pattern variables in the
term being bound. Most of the time this would result in an error at constr
interpretation time. This is not tested anyway in the test-suite, let us try
with the CI.